+++ src/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useAccountInformation';
 export * from './useCryptoPricing';
 export * from './usePasskey';
 export * from './useEventEmitter';
+export * from './useTransactions';
